backend/requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.5.0
django-environ==0.4.5
djangorestframework==3.12.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
djangorestframework-jwt
pytils==0.3